move ScreenshotUtils file under utils
